Rename package model to data
