+++ app/src/main/java/com/jorgesm/themoviedb/data/Error.kt
@@ -1,4 +1,4 @@
-package com.jorgesm.themoviedb.model
+package com.jorgesm.themoviedb.data
 
 import okio.IOException
 import retrofit2.HttpException
