+++ app/src/main/java/com/jorgesm/themoviedb/data/LocationDataSource.kt
@@ -1,7 +1,6 @@
-package com.jorgesm.themoviedb.model
+package com.jorgesm.themoviedb.data
 
 import android.annotation.SuppressLint
-import android.app.Activity
 import android.app.Application
 import android.location.Location
 import com.google.android.gms.location.LocationServices
