+++ app/src/main/java/com/jorgesm/themoviedb/data/RegionRepository.kt
@@ -1,4 +1,4 @@
-package com.jorgesm.themoviedb.model
+package com.jorgesm.themoviedb.data
 
 
 import android.Manifest
