+++ app/src/main/java/com/jorgesm/themoviedb/data/RemoteResult.kt
@@ -1,4 +1,4 @@
-package com.jorgesm.themoviedb.model
+package com.jorgesm.themoviedb.data
 
 
 import com.google.gson.annotations.SerializedName
