+++ app/src/main/java/com/jorgesm/themoviedb/ui/detail/MovieDetailTextInfoView.kt
@@ -5,7 +5,7 @@ import android.util.AttributeSet
 import androidx.appcompat.widget.AppCompatTextView
 import androidx.core.text.bold
 import androidx.core.text.buildSpannedString
-import com.jorgesm.themoviedb.model.database.Movie
+import com.jorgesm.themoviedb.data.database.Movie
 
 class MovieDetailTextInfoView  @JvmOverloads constructor(
     context: Context,
